fix: strip indendtation on code paste
- when pasting code, dedent to normalize the base indendtation to 0, maintaining the relative indentation

Fixes a common issue when pasting nested code from an editor, where e.g. all pasted code is indendent an extra 4, 8, 12 spaces (etc). Group shift+tab is also needed to make the end experience ideal, but this resolves the most common issue
