+++ src/views/edit/editor/plugins/createCodeBlockNormalizationPlugin.tsx
@@ -1,14 +1,20 @@
 import { ELEMENT_CODE_BLOCK } from "@udecode/plate";
-import { createPluginFactory } from "@udecode/plate-common";
+import { createPluginFactory, PlateEditor } from "@udecode/plate-common";
 import { Editor, Transforms } from "slate";
+import { dedent } from "../../../../dedent.js";
+
+const isInCodeBlock = (editor: PlateEditor) => {
+  return Editor.above(editor as Editor, {
+    match: (n: any) => n.type === ELEMENT_CODE_BLOCK,
+  });
+};
 
 /**
- * This plugin handles pasted code to ensure it is pasted as a single block. Prior to this pasting
- * a block of code into the editor would generate a new code_block for each line. This solution
- * was not extensively tested!
+ * This plugin handles pasted code to ensure it is pasted as a single block and indentation
+ * is normalized.
  */
 export const createCodeBlockNormalizationPlugin = createPluginFactory({
-  key: "customPlugin",
+  key: "createCodeBlockNormalizationPlugin",
   withOverrides: (editor) => {
     // todo: types -- Various methods below do `editor as Editor` because editor here is PlateEditor.
     // Unsure what the right way to cast this is -- Grepped a few examples from Plate's codebase
@@ -18,24 +24,25 @@ export const createCodeBlockNormalizationPlugin = createPluginFactory({
     const { insertData } = editor;
 
     editor.insertData = (data: DataTransfer) => {
-      const text = data.getData("text/plain");
-      if (
-        !text ||
-        !Editor.above(editor as Editor, {
-          match: (n: any) => n.type === ELEMENT_CODE_BLOCK,
-        })
-      ) {
+      let text = data.getData("text/plain");
+
+      // Bail if we aren't pasting text into a code block.
+      if (!text || !isInCodeBlock(editor)) {
         insertData(data);
         return;
       }
 
-      const lines = text.split("\n");
       const { selection } = editor;
-      if (selection) {
-        Transforms.insertText(editor as Editor, lines.join("\n"), {
-          at: selection,
-        });
-      }
+      if (!selection) return;
+
+      // strip base indentation: common when pasting from an editor
+      // note this maintains relative indentation, just not the overall
+      // indentation of the block (e.g. copying from a very nested function)
+      text = dedent(text.trim());
+
+      Transforms.insertText(editor as Editor, text, {
+        at: selection,
+      });
     };
 
     return editor;
